+++ src/libmerc/ldap.hpp
@@ -14,6 +14,7 @@
 #include "x509.h"
 #include "json_object.h"
 #include "match.h"
+#include "protocol.h"
 
 namespace ldap {